(refactor) modularize reviewCount management
