+++ js/detail/reviewCount.js
@@ -1,0 +1,10 @@
+export let reviewCount = 0;
+export function incrementReviewCount() {
+  reviewCount++;
+}
+export function decrementReviewCount() {
+  reviewCount--;
+}
+export function resetReviewCount() {
+  reviewCount = 0;
+}
